scripts/chapter7/pyproject.toml
[project]
name = "genai-book-chapter7"
version = "1.0.0"
description = "第7章 AI Agent 開発編のサンプルコード"
requires-python = ">=3.13,<3.14"
dependencies = [
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.29",
    "langchain-tavily>=0.1.1",
    "langgraph>=0.2.76",
    "tavily-python>=0.7.12",
    "python-dotenv>=1.0.1",
    "beautifulsoup4>=4.13.0",
    "requests>=2.32.3",
    "chromadb>=0.5.5",
    "langchain-mcp-adapters>=0.1.0",
]
